Update README.md
Removed instructions to run the app in a virtual environment (GitPod and Codespaces) since this repository can't be opened in virtual environments. 

Corrected the URL in the `git clone` instruction. 

Added a note to tell learners the instructions assume they're using macOS or Linux and to use WSL if they're using Windows.